+++ registration/static/registration.css
@@ -1,1 +1,2 @@
 ul,li{list-style:none}.data-form .form-control{width:100%;max-width:500px}.data-form .radio-group{overflow:hidden}.data-form .save-button{background:#00549F;margin:30px 0}.form-group .subfield input{width:20%}.form-group .subfield .suffix{font-weight:bold}.form-group.postcode input{width:8em;max-width:100%}.form-group.contact-number input{width:11em;max-width:100%}.list-inset{list-style-type:circle;margin-left:2.5em}.grid-row.grid-row--great{background:#F2EEE8;margin-top:30px}.panel--great{background:#F2EEE8;margin-top:30px;padding:1px 15px}.exporting-logo{margin-bottom:30px}.clearfix:after{visibility:hidden;display:block;font-size:0;content:" ";clear:both;height:0}* html .clearfix{zoom:1}*:first-child+html .clearfix{zoom:1}@media (max-width: 979px){body{padding:0}}body h1{font-size:38px}body h2{font-size:26px}body .a11y{border:0;clip:rect(0 0 0 0);height:1px;margin:-1px;overflow:hidden;padding:0;position:absolute;width:1px}body #content{padding-bottom:0}body #content .container a.teal{color:#007ea3}body #content .container a.teal:hover{color:#33cfff}body #content .container .row-fluid .registration-intro-container{margin-top:140px}body #content .container .row-fluid .registration-intro-container .registration-intro{background:rgba(0,0,0,0.6);-ms-filter:progid:DXImageTransform.Microsoft.gradient(startColorstr=#99000000,endColorstr=#99000000);filter:progid:DXImageTransform.Microsoft.gradient(startColorstr=#99000000,endColorstr=#99000000);zoom:1;padding:20px}body #content .container .row-fluid .registration-intro-container .registration-intro .great-britain-logo{margin:0 0 20px;max-width:100px}body #content .container .row-fluid .registration-intro-container .registration-intro .great-britain-logo img{width:100%;height:auto}body #content .container .row-fluid form{margin-bottom:0}body #content .container .row-fluid form .connections-platform-form{background:rgba(0,0,0,0.6);-ms-filter:progid:DXImageTransform.Microsoft.gradient(startColorstr=#99000000,endColorstr=#99000000);filter:progid:DXImageTransform.Microsoft.gradient(startColorstr=#99000000,endColorstr=#99000000);zoom:1;padding:20px 40px 150px}body #content .container .row-fluid form .connections-platform-form h2{margin:40px 0 60px 0}body #content .container .row-fluid form .connections-platform-form .already-registered{margin:0 auto;text-align:center}body #content .container .row-fluid form .connections-platform-form .input-field-container label{margin-bottom:10px}body #content .container .row-fluid form .connections-platform-form .input-field-container input,body #content .container .row-fluid form .connections-platform-form .input-field-container select{border:3px solid #ffffff;border-radius:0;font-size:16px;height:30px;margin:0 0 40px;padding:5px 10px;width:calc(100% - 26px)}body #content .container .row-fluid form .connections-platform-form .input-field-container input:focus,body #content .container .row-fluid form .connections-platform-form .input-field-container select:focus{border:3px solid #ffbf47}body #content .container .row-fluid form .connections-platform-form .input-field-container select{height:50px;width:100%}body #content .container .row-fluid form .connections-platform-form .input-field-container ul.errorlist{margin:0;padding:0}body #content .container .row-fluid form .connections-platform-form .input-field-container ul.errorlist li{color:#ffbf47}body #content .container .row-fluid form .connections-platform-form .register-btn-container{margin:20px 0;text-align:center}body #content .container .row-fluid form .connections-platform-form .register-btn-container input.register-btn{background:#888888;border:0;color:#ffffff;height:60px;width:150px}body #content .container .row-fluid form .connections-platform-form .register-btn-container input.register-btn:hover{background:#666666}body #content .container .row-fluid form .connections-platform-form .steps{margin:20px auto;text-align:center;width:100%;color:white}body #content .container .row-fluid form .connections-platform-form .steps span{font-size:3em}body #content .container .registered,body #content .container .sorry{background:rgba(0,0,0,0.6);-ms-filter:progid:DXImageTransform.Microsoft.gradient(startColorstr=#99000000,endColorstr=#99000000);filter:progid:DXImageTransform.Microsoft.gradient(startColorstr=#99000000,endColorstr=#99000000);zoom:1;margin-bottom:150px;margin-top:150px;padding:20px}body #content .container .registered h1,body #content .container .registered h2,body #content .container .sorry h1,body #content .container .sorry h2{margin-bottom:20px}body{background:url(images/banner-img1.jpg) no-repeat center center fixed;-webkit-background-size:cover;-moz-background-size:cover;-o-background-size:cover;background-size:cover}.container{color:#ffffff}@media (max-width: 992px){body #content{margin:0 20px}body #content .container{width:100%}body #content .container .row-fluid form .connections-platform-form-container{width:100%}body #content .container .row-fluid form .connections-platform-form-container .connections-platform-form{height:100%;margin:40px 0 0;padding:0 20px}body #content .container .row-fluid form .connections-platform-form-container .connections-platform-form .register-btn-container input.register-btn{width:50%}}@media (max-width: 767px){body #content .container .row-fluid .registration-intro-container{margin-top:30px;margin-bottom:30px}body #content .container .row-fluid form .connections-platform-form{padding-bottom:50px}body #content .container .registered,body #content .container .sorry{margin-top:30px;margin-bottom:30px}}
+/*# sourceMappingURL=registration.css.map */
