+++ user/static/sass/user-profile-details.scss
