+++ user/tests/test_templates.py
